fix(quiz): Fix incorrect references to long when it should have been integer [SCRUM-201]

+++ backend/src/main/java/stud/ntnu/backend/service/QuizService.java
@@ -73,7 +73,7 @@ public class QuizService {
     quizRepository.save(quiz);
   }
 
-  public void createUserQuizAttempt(Long quizId, Long userId) {
+  public void createUserQuizAttempt(Long quizId, Integer userId) {
     UserQuizAttempt attempt = new UserQuizAttempt();
     attempt.setUserId(userId);
     attempt.setQuizId(quizId);
@@ -90,13 +90,13 @@ public class QuizService {
     userQuizAnswerRepository.save(answer);
   }
 
-  public Page<QuizAttemptSummaryDto> getQuizAttemptsByQuizId(Long quizId, Long userId,
+  public Page<QuizAttemptSummaryDto> getQuizAttemptsByQuizId(Long quizId, Integer userId,
       Pageable pageable) {
     return userQuizAttemptRepository.findByUserIdAndQuizId(userId, quizId, pageable)
         .map(a -> new QuizAttemptSummaryDto(a.getId(), a.getCompletedAt()));
   }
 
-  public Page<QuizBasicInfoDto> getBasicInfoForAttemptedQuizzes(Long userId, Pageable pageable) {
+  public Page<QuizBasicInfoDto> getBasicInfoForAttemptedQuizzes(Integer userId, Pageable pageable) {
     // Find all quiz IDs for which the user has at least one attempt
     List<UserQuizAttempt> attempts = userQuizAttemptRepository.findByUserId(userId);
     Set<Long> quizIds = attempts.stream().map(UserQuizAttempt::getQuizId)
